requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.0
Pillow>=10.0.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
